REFACTOR AND RESTYLE : client side
refactoring : add custom hooks
restyle : add disabling btn after the btn clicked
