+++ projects/client/src/components/NavbarAdmin.jsx
@@ -1,51 +1,54 @@
-import React, { useState } from "react";
+import React from "react";
 import NavbarRenderIcon from "./NavbarRenderIcon";
-import { useNavigate } from "react-router-dom";
-import { useDispatch, useSelector } from "react-redux";
-import { setLoggedInAdminDataBack } from "../feature/admin_auth/slice/AdminLogInSlice";
+import { useAdminNavbar } from "../utils/useAdminNavbar";
+import HamburgToggle from "./HamburgToggle";
 
 function NavbarAdmin() {
-  const [toggleNav, setToggleNav] = useState(false);
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const admin = useSelector((state) => state.adminLogin.loggedInAdminData);
+  const { toggleNav, setToggleNav, navigate, admin, listNav, logoutBtnHandler, onClickNavbarBtn } = useAdminNavbar();
 
-  const listNav = [
-    {
-      class: "uil uil-cog",
-      text: "user",
-      navlink: "/admin/dashboard/user-management",
-    },
-    {
-      class: "uil uil-estate",
-      text: "warehouse",
-      navlink: "/admin/dashboard/warehouse-management",
-    },
-    {
-      class: "uil uil-suitcase",
-      text: "prod. & cat.",
-      navlink: "/admin/dashboard/product-management",
-    },
-    {
-      class: "uil uil-file-graph",
-      text: "prod-order",
-      navlink: "/admin/dashboard/order",
-    },
-    {
-      class: "uil uil-analytics",
-      text: "report",
-      navlink: "/admin/dashboard/report",
-    },
-  ];
-
-  const logoutBtnHandler = () => {
-    localStorage.removeItem("admin_token");
-    dispatch(setLoggedInAdminDataBack());
-    navigate("/");
+  const NavbarTitle = () => {
+    return (
+      <div
+        onClick={() => navigate("/admin/dashboard")}
+        className="invisible md:visible md:text-lg text-primary
+        md:pt-4 md:pb-4 scale-y-110 font-bold lg:text-2xl hover:cursor-pointer font-title text-center"
+      >
+        <h1>warehouse</h1>
+        <h1 className="font-bold font-body text-sm">{admin.username}</h1>
+      </div>
+    );
   };
 
-  const onClickNavbarBtn = async (navItem) => {
-    navigate(navItem.navlink);
+  const NavbarIcons = () => {
+    return (
+      <ul
+        id="navbar-items"
+        className="grid grid-cols-3 gap-4 text-lg text-primary md:grid 
+        md:grid-rows-6 md:grid-cols-1 md:gap-0 md:text-2xl py-4 lg:text-3xl"
+      >
+        <NavbarRenderIcon listNav={listNav} onClickNavbarBtn={onClickNavbarBtn} />
+        <li
+          className="text-center md:py-4 text-red-600 hover:text-white hover:bg-red-600 hover:cursor-pointer"
+          onClick={logoutBtnHandler}
+        >
+          <i className="uil uil-signout"></i>
+          <h2 className="text-base md:text-lg lg:text-xl">logout</h2>
+        </li>
+      </ul>
+    );
+  };
+
+  const ClosedBtnNavbarAdmin = () => {
+    return (
+      <span
+        className="md:invisible absolute top-2 right-0 text-red-900 text-sm hover:cursor-pointer"
+        onClick={() => {
+          setToggleNav(false);
+        }}
+      >
+        <i className="uil uil-times-circle"></i>
+      </span>
+    );
   };
 
   return (
@@ -62,56 +65,12 @@ function NavbarAdmin() {
         }
       >
         <div className="relative">
-          <div
-            onClick={() => navigate("/admin/dashboard")}
-            className="invisible md:visible md:text-lg text-primary
-            md:pt-4 md:pb-4 scale-y-110 font-bold lg:text-2xl hover:cursor-pointer font-title text-center"
-          >
-            <h1>warehouse</h1>
-            <h1 className="font-bold font-body text-sm">{admin.username}</h1>
-          </div>
-          <ul
-            id="navbar-items"
-            className="grid grid-cols-3 gap-4 text-lg text-primary md:grid 
-            md:grid-rows-6 md:grid-cols-1 md:gap-0 md:text-2xl py-4 lg:text-3xl"
-          >
-            <NavbarRenderIcon listNav={listNav} onClickNavbarBtn={onClickNavbarBtn} />
-            <li
-              className="text-center md:py-4 text-red-600 hover:text-white hover:bg-red-600 hover:cursor-pointer"
-              onClick={logoutBtnHandler}
-            >
-              <i className="uil uil-signout"></i>
-              <h2 className="text-base md:text-lg lg:text-xl">logout</h2>
-            </li>
-          </ul>
-          <span
-            className="md:invisible absolute top-2 right-0 text-red-900 text-sm hover:cursor-pointer"
-            onClick={() => {
-              setToggleNav(false);
-            }}
-          >
-            <i className="uil uil-times-circle"></i>
-          </span>
+          <NavbarTitle />
+          <NavbarIcons />
+          <ClosedBtnNavbarAdmin />
         </div>
       </div>
-      <div
-        className={
-          toggleNav
-            ? `invisible absolute bottom-2 right-2 text-2xl text-center
-            md:invisible hover:cursor-pointer z-0 h-10 w-10 bg-primary
-             self-center text-white rounded-full flex flex-row justify-center
-              items-center`
-            : `visible absolute bottom-2 right-2 text-2xl text-center
-            md:invisible hover:cursor-pointer z-0 h-10 w-10 bg-primary
-             self-center text-white rounded-full flex flex-row justify-center
-              items-center`
-        }
-        onClick={() => {
-          setToggleNav(true);
-        }}
-      >
-        <i className="uil uil-align-center-alt"></i>
-      </div>
+      <HamburgToggle toggleNav={toggleNav} setToggleNav={setToggleNav} />
     </>
   );
 }
