+++ projects/client/src/feature/admin/component/all_admin/create_data/AddNewAdmin.jsx
@@ -1,54 +1,15 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { Formik, Form } from "formik";
-import * as Yup from "yup";
 import CustomForm from "../../../../../components/CustomForm";
 import CustomSelect from "../../../../../components/CustomSelect";
-import { useDispatch } from "react-redux";
 import RenderCity from "../edit_data/RenderCity";
 import RenderWarehouse from "../edit_data/RenderWarehouse";
-import { createNewAdmin, getAllAdmin, getWarehousesInCities } from "../../../";
 import ClosedBtnModal from "../../../../../components/ClosedBtnModal";
+import { useAddAdmin } from "../../../util/useAddAdmin";
 
 function AddNewAdmin(props) {
   const { setNewAdminClicked, warehouseCities, page } = props;
-  const dispatch = useDispatch();
-  const [selectCity, setCity] = useState();
-  const [warehouses, setWarehouses] = useState([]);
-
-  // get warehouse list after select city
-  const getDataWarehouse = async (cityInput) => {
-    const data = await getWarehousesInCities(cityInput);
-    setWarehouses([...data]);
-  };
-
-  useEffect(() => {
-    getDataWarehouse(selectCity);
-  }, [selectCity]);
-  // =========================================
-
-  const phoneRegExp =
-    /^((\\+[1-9]{1,4}[ \\-]*)|(\\([0-9]{2,3}\\)[ \\-]*)|([0-9]{2,4})[ \\-]*)*?[0-9]{3,4}?[ \\-]*[0-9]{3,4}?$/;
-  const passwordRegex = /^(?=.*\d).{6,}$/;
-
-  const registerSchema = Yup.object().shape({
-    username: Yup.string().required("must not blank"),
-    email: Yup.string().required("must not blank").email("invalid email format"),
-    password: Yup.string()
-      .min(6, "password is too short - at least 6 chars minimum")
-      .matches(passwordRegex, "must contain 1 number")
-      .required(),
-    confirmPassword: Yup.string().oneOf([Yup.ref("password"), null], "Passwords must match"),
-    phone_number: Yup.string().required("must not blank").matches(phoneRegExp, "phone number is not valid"),
-    id_city: Yup.number("required").required("required"),
-    id_warehouse: Yup.number("required").required("required"),
-  });
-
-  const onSubmit = async (values, action) => {
-    const result = await createNewAdmin(values);
-    alert(result.message);
-    await dispatch(getAllAdmin(page));
-    setNewAdminClicked(false);
-  };
+  const { setCity, warehouses, registerSchema, onSubmit } = useAddAdmin(setNewAdminClicked, page);
 
   return (
     <div className="modal-background">
@@ -96,7 +57,7 @@ function AddNewAdmin(props) {
                   </CustomSelect>
                   <div className="row-start-8 row-span-1">
                     <div className="grid grid-cols-2 gap-8 text-sm h-5/6 mt-4">
-                      <button className="bg-primary text-white" type="submit">
+                      <button id="create-admin-btn" className="bg-primary text-white btn-disabled" type="submit">
                         Create
                       </button>
                     </div>
