+++ projects/client/src/feature/admin/component/all_admin/edit_data/ConfirmationModal.jsx
@@ -7,7 +7,12 @@ function ConfirmationModal(props) {
       <div className="conf-modal-container">
         <h1 className="conf-modal-header-text">Are you sure to submit?</h1>
         <div className="conf-modal-btn-container">
-          <button className="conf-modal-btn-confirm" type="submit" disabled={formikProps.isSubmitting}>
+          <button
+            id="confirm-admin-btn"
+            className="conf-modal-btn-confirm btn-disabled"
+            type="submit"
+            disabled={formikProps.isSubmitting}
+          >
             Confirm
           </button>
           <div className="conf-modal-btn-cancel" onClick={() => props.setConfirmationModal(false)}>
