+++ projects/client/src/feature/admin_product/component/categories/edit_category/EditModal.jsx
@@ -38,7 +38,12 @@ function EditModal(props) {
               ) : null}
             </div>
             <div className="grid grid-cols-3 gap-2 text-sm h-8">
-              <button type="submit" onClick={formik.handleSubmit} className="bg-primary text-white h-full">
+              <button
+                id="edit-category-btn"
+                type="submit"
+                onClick={formik.handleSubmit}
+                className="bg-primary text-white h-full btn-disabled"
+              >
                 Submit
               </button>
             </div>
