+++ projects/client/src/feature/admin_product/component/manage_stock/BodyStock.jsx
@@ -19,43 +19,53 @@ function BodyStock(props) {
     deleteBtnHandler,
   } = useStockBody(userAdmin);
 
+  const ManageStockModals = () => {
+    return (
+      <>
+        {isEditClicked ? (
+          <EditModal
+            setEditClicked={setEditClicked}
+            warehouses={warehouses}
+            productStock={productStock}
+            userAdmin={userAdmin}
+            singleProduct={singleProduct}
+            refetchedData={refetchedData}
+          />
+        ) : null}
+        {isDeleteClicked ? (
+          <DeleteModal
+            setDeleteClicked={setDeleteClicked}
+            warehouses={warehouses}
+            productStock={productStock}
+            userAdmin={userAdmin}
+            singleProduct={singleProduct}
+            refetchedData={refetchedData}
+          />
+        ) : null}
+      </>
+    );
+  };
+
+  const ManageStockDataBody = () => {
+    return (
+      <div className="row-span-11 grid grid-rows-10 gap-2 lg:gap-2">
+        <div className="row-span-1 font-semibold grid grid-cols-6 items-end text-xs pl-2 md:text-sm lg:text-base">
+          <p className="col-span-2">name</p>
+          <p className="hidden lg:inline lg:col-span-1">weight (kg)</p>
+          <p className="col-span-1  lg:col-span-1 text-center ">stock</p>
+          <p className="col-span-2 lg:col-span-1 text-center">booked qty</p>
+          <p className="text-right">action</p>
+        </div>
+        <RenderProduct productsList={productsList} editBtnHndler={editBtnHndler} deleteBtnHandler={deleteBtnHandler} />
+      </div>
+    );
+  };
+
   return (
     <>
-      {isEditClicked ? (
-        <EditModal
-          setEditClicked={setEditClicked}
-          warehouses={warehouses}
-          productStock={productStock}
-          userAdmin={userAdmin}
-          singleProduct={singleProduct}
-          refetchedData={refetchedData}
-        />
-      ) : null}
-      {isDeleteClicked ? (
-        <DeleteModal
-          setDeleteClicked={setDeleteClicked}
-          warehouses={warehouses}
-          productStock={productStock}
-          userAdmin={userAdmin}
-          singleProduct={singleProduct}
-          refetchedData={refetchedData}
-        />
-      ) : null}
+      <ManageStockModals />
       <div className="row-span-6 grid grid-rows-12">
-        <div className="row-span-11 grid grid-rows-10 gap-2 lg:gap-2">
-          <div className="row-span-1 font-semibold grid grid-cols-6 items-end text-xs pl-2 md:text-sm lg:text-base">
-            <p className="col-span-2">name</p>
-            <p className="hidden lg:inline lg:col-span-1">weight (kg)</p>
-            <p className="col-span-1  lg:col-span-1 text-center ">stock</p>
-            <p className="col-span-2 lg:col-span-1 text-center">booked qty</p>
-            <p className="text-right">action</p>
-          </div>
-          <RenderProduct
-            productsList={productsList}
-            editBtnHndler={editBtnHndler}
-            deleteBtnHandler={deleteBtnHandler}
-          />
-        </div>
+        <ManageStockDataBody />
         <div className="pagination-container">
           <AdminPagination setPageNum={setPageNum} pageNum={pageNum} totalPage={totalPage} />
         </div>
