+++ projects/client/src/feature/admin_product/component/manage_stock/edit_data/SubmitButton.jsx
@@ -9,6 +9,7 @@ function SubmitButton(props) {
 
   return (
     <button
+      id="edit-stock-btn"
       disabled={noWarehouseSelected()}
       className="bg-primary text-white h-full disabled:bg-slate-300 disabled:hover:cursor-not-allowed"
       onClick={submitBtnHandler}
