+++ projects/client/src/feature/admin_product/component/product/add_data/AddDataModal.jsx
@@ -41,7 +41,12 @@ function AddDataModal(props) {
               <RenderCategoryOptions categories={categories} />
             </CustomSelectFormikHook>
             <div className="grid grid-cols-3 gap-2 text-sm h-8 mt-4">
-              <button type="submit" onClick={formik.handleSubmit} className="bg-primary text-white h-full">
+              <button
+                id="add-product-btn"
+                type="submit"
+                onClick={formik.handleSubmit}
+                className="bg-primary text-white h-full btn-disabled"
+              >
                 Submit
               </button>
             </div>
