+++ projects/client/src/feature/admin_product/component/product/delete_data/DeleteModal.jsx
@@ -9,9 +9,12 @@ function DeleteModal(props) {
   const [preview, setPreview] = useState(`${REACT_APP_SERVER_URL + singleProduct.product_image}`);
 
   const deleteDataHandler = async () => {
+    const deleteProductBtn = document.getElementsByClassName("delete-modal-confirmation-btn");
+    deleteProductBtn.disabled = true;
     const response = await deleteProduct(singleProduct.id_product);
     alert(response.message);
     await refetchedData();
+    deleteProductBtn.disabled = false;
     setDeleteClicked(false);
   };
 
@@ -41,7 +44,7 @@ function DeleteModal(props) {
             Are you sure want to <i className="font-bold">delete</i> this data?
           </h3>
           <div className="w-full flex justify-center">
-            <button className="delete-modal-confirmation-btn" onClick={deleteDataHandler}>
+            <button className="delete-modal-confirmation-btn btn-disabled" onClick={deleteDataHandler}>
               Delete
             </button>
           </div>
