+++ projects/client/src/feature/admin_product/component/product/edit_data/EditModal.jsx
@@ -78,6 +78,7 @@ function EditModal(props) {
             </CustomSelectFormikHook>
             <div className="grid grid-cols-3 gap-2 text-sm h-8 mt-4">
               <button
+                id="edit-product-btn"
                 type="submit"
                 onClick={formik.handleSubmit}
                 className="bg-primary text-white h-full disabled:cursor-not-allowed
