+++ projects/client/src/feature/admin_warehouse/component/edit_data/EditModal.jsx
@@ -52,7 +52,7 @@ function EditModal(props) {
                   <div className="invisible"></div>
                   <div className=" row-span-1">
                     <div className="grid grid-cols-2 gap-8 text-sm h-5/6 mt-4">
-                      <button className="bg-primary text-white" type="submit">
+                      <button id="edit-warehouse-btn" className="bg-primary text-white btn-disabled" type="submit">
                         Submit
                       </button>
                     </div>
