+++ projects/client/src/feature/stock_mutation/component/order_management/OrderManagementBody.jsx
@@ -1,69 +1,46 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import SelectFilter from "../../../../components/SelectFilter";
 import RenderWarehouse from "../../../admin/component/all_admin/edit_data/RenderWarehouse";
 import NoData from "../../../../components/NoData";
 import RenderBodyData from "./RenderBodyData";
 import AdminPagination from "../../../../components/AdminPagination";
-import { getOrderProducts } from "../../";
 import RenderOrderProducts from "./RenderOrderProducts";
 import DetailOrder from "./DetailOrder";
+import { useOrderMgtBody } from "../../util/useOrderMgtBody";
 
 function OrderManagementBody(props) {
   const { admin, warehouses } = props;
-  const [orderData, setOrderData] = useState([]);
-  const [singleOrder, setSingleOrder] = useState({});
-  const [isSingleItemClicked, setSingleItemClicked] = useState(false);
-  const [pageNum, setPageNum] = useState(1);
-  const [totalPage, setTotalPage] = useState(1);
-  const [filterState, setFilterState] = useState({
-    warehouse: admin.id_warehouse ? admin.id_warehouse : "",
-    status: "",
-  });
-  const OFFSET = 7;
-  const LIMIT = 7;
+  const {
+    orderData,
+    singleOrder,
+    isSingleItemClicked,
+    setSingleItemClicked,
+    pageNum,
+    setPageNum,
+    totalPage,
+    filterState,
+    fetchingData,
+    isWarehouseFilterDisabled,
+    selectFilterOnChange,
+    singleItemHandler,
+  } = useOrderMgtBody(admin);
 
-  const fetchingData = async () => {
-    const dataInput = {
-      offset: OFFSET,
-      limit: LIMIT,
-      page: pageNum,
-      id_warehouse: filterState.warehouse,
-      status_order: filterState.status,
-    };
-    const response = await getOrderProducts(dataInput);
-    setOrderData([...response?.result?.dataToSend]);
-    setTotalPage(response?.result?.totalPage);
+  const OrderMgtModals = () => {
+    return (
+      <>
+        {isSingleItemClicked ? (
+          <DetailOrder
+            singleOrder={singleOrder}
+            setSingleItemClicked={setSingleItemClicked}
+            fetchingData={fetchingData}
+          />
+        ) : null}
+      </>
+    );
   };
 
-  const isWarehouseFilterDisabled = () => {
-    return admin.id_role !== 1;
-  };
-
-  const selectFilterOnChange = (e) => {
-    const subFilterName = e.target.id.split("-")[0];
-    const filterChange = { [subFilterName]: e.target.value };
-    setFilterState((prevState) => ({ ...prevState, ...filterChange }));
-    setPageNum(1);
-  };
-
-  const singleItemHandler = async (singleData) => {
-    setSingleItemClicked(true);
-    setSingleOrder({ ...singleData });
-  };
-
-  useEffect(() => {
-    fetchingData();
-  }, [filterState, pageNum]);
-
-  return (
-    <>
-      {isSingleItemClicked ? (
-        <DetailOrder
-          singleOrder={singleOrder}
-          setSingleItemClicked={setSingleItemClicked}
-          fetchingData={fetchingData}
-        />
-      ) : null}
+  const OrderMgtFilters = () => {
+    return (
       <form className="grid grid-cols-3 gap-4 md:gap-4 md:grid-cols-4 lg:grid-cols-5 text-xs md:text-sm lg:text-base h-4/5">
         <SelectFilter
           text="warehouse"
@@ -83,6 +60,13 @@ function OrderManagementBody(props) {
           <option value={"canceled"}>canceled</option>
         </SelectFilter>
       </form>
+    );
+  };
+
+  return (
+    <>
+      <OrderMgtModals />
+      <OrderMgtFilters />
       <div className="row-span-10 grid grid-rows-8 gap-2 lg:gap-2">
         {orderData?.length ? (
           <RenderBodyData>
