+++ projects/client/src/feature/stock_mutation/component/stock_mutation/update_mutation/UpdateModal.jsx
@@ -3,25 +3,17 @@ import ClosedBtnModal from "../../../../../components/ClosedBtnModal";
 import FieldDataRender from "./FieldDataRender";
 import RequestedAction from "./RequestedAction";
 import RequesterAction from "./RequesterAction";
-import { acceptDeliveredProducts, approveMutationRequests, rejectMutationRequests } from "../../../";
+import { useUpdateMutation } from "../../../util/useUpdateMutation";
 
 function UpdateModal(props) {
   const { setSingleItemClicked, admin, singleData, fetchingData } = props;
-  console.log(singleData);
 
-  const statusMutation = () => {
-    let status;
-    if (!singleData.is_approve && !singleData.is_reject) {
-      status = "waiting for approval";
-    } else if (singleData.is_accepted) {
-      status = "shipped";
-    } else if (singleData.is_reject) {
-      status = "rejected";
-    } else if (singleData.is_sending) {
-      status = "on-delivery";
-    }
-    return status;
-  };
+  const { acceptBtnHandler, rejectBtnHandler, approveBtnHandler, dateFormatting, statusMutation } = useUpdateMutation(
+    setSingleItemClicked,
+    admin,
+    singleData,
+    fetchingData,
+  );
 
   const LastUpdatedBy = () => {
     if (singleData.is_accepted) {
@@ -31,46 +23,6 @@ function UpdateModal(props) {
     } else if (singleData.is_approve) {
       return <FieldDataRender textLabel="approved and sent at" value={dateFormatting(singleData.updatedAt)} />;
     } else return <></>;
-  };
-
-  const dateFormatting = (dateString) => {
-    const date = new Date(dateString);
-    const formattedDateTime = date.toLocaleString("en-US", {
-      month: "numeric",
-      day: "numeric",
-      year: "numeric",
-      hour: "numeric",
-      minute: "numeric",
-      second: "numeric",
-      hour12: true,
-      timeZone: "UTC", // Adjust this according to your desired time zone
-    });
-    return `${formattedDateTime}`; // Combine date and time
-  };
-
-  const approveBtnHandler = async () => {
-    const data = { adminData: admin, mutationData: singleData };
-    console.log(JSON.stringify(data));
-    const response = await approveMutationRequests(data);
-    alert(response.message);
-    await fetchingData();
-    setSingleItemClicked(false);
-  };
-
-  const rejectBtnHandler = async () => {
-    const data = { adminData: admin, mutationData: singleData };
-    const response = await rejectMutationRequests(data);
-    alert(response.message);
-    await fetchingData();
-    setSingleItemClicked(false);
-  };
-
-  const acceptBtnHandler = async () => {
-    const data = { adminData: admin, mutationData: singleData };
-    const response = await acceptDeliveredProducts(data);
-    alert(response.message);
-    await fetchingData();
-    setSingleItemClicked(false);
   };
 
   return (
